Fix duplicate dependency, update gradle and change object to extension
