+++ app/build.gradle
@@ -45,13 +45,12 @@ ext {
     ankoLibVersion = "0.10.1"
     supportLibVersion = '26.0.1'
     gsonVersion = '2.8.0'
-    retrofitVersion = '2.2.0'
     rxKotlinVersion = '2.1.0'
     rxAndroidVersion = '2.0.1'
     retrofitVersion = '2.3.0'
     glideVersion = '4.0.0'
     flexboxVersion = '0.3.0-alpha4'
-    archRoomVersion = '1.0.0-alpha8'
+    archRoomVersion = '1.0.0-alpha9'
     daggerVersion = '2.11'
     deeplinkDispatchVersion = "3.1.0"
     junitVersion = "4.12"
