+++ app/src/main/java/io/github/tonnyl/mango/ui/shot/ShotFragment.kt
@@ -18,7 +18,9 @@ import android.widget.TextView
 import io.github.tonnyl.mango.R
 import io.github.tonnyl.mango.data.Shot
 import io.github.tonnyl.mango.data.User
-import io.github.tonnyl.mango.glide.GlideLoader
+import io.github.tonnyl.mango.extension.OnPaletteProcessCallback
+import io.github.tonnyl.mango.extension.loadAvatar
+import io.github.tonnyl.mango.extension.loadHighQualityWithPalette
 import io.github.tonnyl.mango.ui.shot.comments.CommentsActivity
 import io.github.tonnyl.mango.ui.shot.comments.CommentsPresenter
 import io.github.tonnyl.mango.ui.shot.likes.LikesActivity
@@ -111,7 +113,7 @@ class ShotFragment : Fragment(), ShotContract.View {
 
     override fun show(shot: Shot) {
         // Show the menu_shot image
-        GlideLoader.loadHighQualityWithPalette(context, shot_image_view, shot.images.best(), object : GlideLoader.OnPaletteProcessCallback {
+        shot_image_view.loadHighQualityWithPalette(shot.images.best(), object : OnPaletteProcessCallback {
             override fun OnPaletteGenerated(palette: Palette?) {
                 palette?.let {
                     showPalette(palette)
@@ -128,7 +130,7 @@ class ShotFragment : Fragment(), ShotContract.View {
         })
 
         shot.user?.let {
-            GlideLoader.loadAvatar(context, avatar, it.avatarUrl)
+            avatar.loadAvatar(it.avatarUrl)
         }
 
         toolbar.title = shot.title
