+++ app/src/main/java/io/github/tonnyl/mango/ui/shot/comments/CommentsAdapter.kt
@@ -10,7 +10,7 @@ import android.view.View
 import android.view.ViewGroup
 import io.github.tonnyl.mango.R
 import io.github.tonnyl.mango.data.Comment
-import io.github.tonnyl.mango.glide.GlideLoader
+import io.github.tonnyl.mango.extension.loadAvatar
 import kotlinx.android.synthetic.main.item_comment.view.*
 
 /**
@@ -30,7 +30,7 @@ class CommentsAdapter(context: Context, list: List<Comment>) : RecyclerView.Adap
     override fun onBindViewHolder(viewHolder: RecyclerView.ViewHolder?, position: Int) {
         val comment = mList[position]
         with(viewHolder as CommentViewHolder) {
-            GlideLoader.loadAvatar(mContext, itemView.avatar, comment.user.avatarUrl)
+            itemView.avatar.loadAvatar(comment.user.avatarUrl)
 
             itemView.comment_body.text = if (Build.VERSION.SDK_INT >= 24) {
                 Html.fromHtml(comment.body, Html.FROM_HTML_MODE_LEGACY)
