+++ app/src/main/java/io/github/tonnyl/mango/ui/shot/likes/LikesAdapter.kt
@@ -7,7 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import io.github.tonnyl.mango.R
 import io.github.tonnyl.mango.data.Like
-import io.github.tonnyl.mango.glide.GlideLoader
+import io.github.tonnyl.mango.extension.loadAvatar
 import kotlinx.android.synthetic.main.item_user.view.*
 
 
@@ -28,7 +28,7 @@ class LikesAdapter(context: Context, list: List<Like>) : RecyclerView.Adapter<Re
     override fun onBindViewHolder(viewHolder: RecyclerView.ViewHolder?, position: Int) {
         val like = mList[position]
         with(viewHolder as LikeViewHolder) {
-            GlideLoader.loadAvatar(mContext, itemView.avatar, like.user.avatarUrl)
+            itemView.avatar.loadAvatar(like.user.avatarUrl)
             itemView.name.text = like.user.name
             itemView.user_name.text = like.user.username
         }
